Fixed IcePHP assert that would occur with a PHP debug executable when destroying multiple times the same communicator (ICE-8635)

+++ php/src/php5/Communicator.cpp
@@ -154,7 +154,7 @@ public:
     {
         _registeredCommunicatorsMutex = new IceUtil::Mutex();
     }
-    
+
     ~Init()
     {
         delete _registeredCommunicatorsMutex;
@@ -182,38 +182,40 @@ ZEND_METHOD(Ice_Communicator, destroy)
     CommunicatorInfoIPtr _this = Wrapper<CommunicatorInfoIPtr>::value(getThis() TSRMLS_CC);
     assert(_this);
 
-    //
-    // Remove all registrations.
-    //
-    {
-        IceUtilInternal::MutexPtrLock<IceUtil::Mutex> lock(_registeredCommunicatorsMutex);
-        for(vector<string>::iterator p = _this->ac->ids.begin(); p != _this->ac->ids.end(); ++p)
-        {
-            _registeredCommunicators.erase(*p);
-        }
-        _this->ac->ids.clear();
-    }
-
-    //
-    // We need to destroy any object factories installed by this request.
-    //
-    _this->destroyObjectFactories(TSRMLS_C);
-
     Ice::CommunicatorPtr c = _this->getCommunicator();
     assert(c);
     CommunicatorMap* m = reinterpret_cast<CommunicatorMap*>(ICE_G(communicatorMap));
     assert(m);
-    assert(m->find(c) != m->end());
-    m->erase(c);
+    if(m->find(c) != m->end()) // If not already destroyed
+    {
+        m->erase(c);
 
-    try
-    {
-        c->destroy();
-    }
-    catch(const IceUtil::Exception& ex)
-    {
-        throwException(ex TSRMLS_CC);
-        RETURN_NULL();
+        //
+        // Remove all registrations.
+        //
+        {
+            IceUtilInternal::MutexPtrLock<IceUtil::Mutex> lock(_registeredCommunicatorsMutex);
+            for(vector<string>::iterator p = _this->ac->ids.begin(); p != _this->ac->ids.end(); ++p)
+            {
+                _registeredCommunicators.erase(*p);
+            }
+            _this->ac->ids.clear();
+        }
+
+        //
+        // We need to destroy any object factories installed by this request.
+        //
+        _this->destroyObjectFactories(TSRMLS_C);
+
+        try
+        {
+            c->destroy();
+        }
+        catch(const IceUtil::Exception& ex)
+        {
+            throwException(ex TSRMLS_CC);
+            RETURN_NULL();
+        }
     }
 }
 
@@ -334,7 +336,7 @@ ZEND_METHOD(Ice_Communicator, proxyToProperty)
                 RETURN_NULL();
             }
             assert(prx);
-            
+
             Ice::PropertyDict val = _this->getCommunicator()->proxyToProperty(prx, prefix);
             if(!createStringMap(return_value, val TSRMLS_CC))
             {
@@ -501,7 +503,7 @@ ZEND_METHOD(Ice_Communicator, getLogger)
     {
         WRONG_PARAM_COUNT;
     }
-    
+
     CommunicatorInfoIPtr _this = Wrapper<CommunicatorInfoIPtr>::value(getThis() TSRMLS_CC);
     assert(_this);
